auth_with_firebase/requirements.txt
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-firebase-auth==1.0.0
